修改标签config

+++ 云渠道/Moudel/工作/二手房/勘察维护/勘察维护详情/MaintainDetailVC.m
@@ -397,8 +397,9 @@
                 if (section == 2) {
                     
                     ModifyTagVC *nextVC = [[ModifyTagVC alloc] initWithArray:_detailDic[@"house_tags"]];
+//                    nextVC.status = [NSString stringWithFormat:@"%ld",_type];
                     nextVC.houseId = _houseId;
-                    nextVC.typeId = [NSString stringWithFormat:@"%ld",_type];
+                    nextVC.typeId = [NSString stringWithFormat:@"%ld",(long)_type];
                     nextVC.modifyTagSaveBtnBlock = ^(NSArray *array) {
                         
                         [self RequestMethod];
